Add files via upload
One of the most common problems facing gyms and other services is customer churn. How do you find out if a customer is no longer with you? Churn indicators vary from field to field. In the case of a gym, it makes sense to say that a customer has left if he or she does not come for a month. In order to combat churn, Model Fitness has digitized several of its customer profiles. The task is to analyze them and develop a customer retention strategy.


Using Machine Learning we will:
 - Learn to predict the probability of churn (for the next month) for each customer.
 - Develop portraits of typical users: select the most prominent groups and describe their main characteristics.
 - Analyze the factors that most impact churn.
 - Draw basic conclusions and develop recommendations to improve customer service:
 - identify target groups;
 - suggest measures to reduce churn;
 - describe any other patterns you observe regarding customer interactions.